chore(symbol): move adapters of Clipboard and Draft to symbol
