+++ app/src/main/java/com/osfans/trime/ime/symbol/LiquidKeyboard.java
@@ -10,11 +10,9 @@ import com.google.android.flexbox.FlexWrap;
 import com.google.android.flexbox.FlexboxLayoutManager;
 import com.google.android.flexbox.JustifyContent;
 import com.osfans.trime.R;
-import com.osfans.trime.clipboard.ClipboardAdapter;
 import com.osfans.trime.data.db.DbBean;
 import com.osfans.trime.data.db.clipboard.ClipboardDao;
 import com.osfans.trime.data.Config;
-import com.osfans.trime.draft.DraftAdapter;
 import com.osfans.trime.data.db.draft.DraftDao;
 import com.osfans.trime.ime.core.Trime;
 import com.osfans.trime.ime.enums.SymbolKeyboardType;
